refactor(convert): split tools into multiple files
