+++ tests/unit/test_tool.py
@@ -5,10 +5,8 @@ from mavis.tools import (
     _convert_tool_row,
     SUPPORTED_TOOL,
     _parse_transabyss,
-    _parse_chimerascan,
-    _parse_bnd_alt,
-    _parse_vcf_record,
 )
+from mavis.tools.vcf import parse_bnd_alt as _parse_bnd_alt, convert_record as _parse_vcf_record
 
 from .mock import Mock
